AndyActivity
Adding a killOnDestroy flag to kill the process when Activity is
Destroyed. This allows better kill option for extending classes

the contentView is set in this class itself. Extending classes must
implement and pass a View via createContentView(). Forces the extending
classes to pass something useful


AndyFace

passing Activity instead of AndyActivity since it is not compulsory for
developers to make their Activity extend AndyActivity. More freedom!